requirements.txt
html5lib==1.1
beautifulsoup4==4.11.2
lxml==4.9.2
requests==2.28.2
selenium-wire==5.1.0
marshmallow==3.19.0
marshmallow_dataclass==8.5.11
google-api-python-client==2.81.0
oauth2client==4.1.3
google_auth_oauthlib==1.0.0
telebot==0.0.5
jinja2==3.1.2
MarkupSafe==2.1.2
aiohttp_jinja2==1.5.1
asyncpg==0.27.0